requirements.txt
typer[all]==0.9.0        
pillow==10.3.0           
pymongo==4.7.2           
rich==13.7.1             
python-dotenv==1.0.1      
